requirements.txt
python-dotenv
requests
PyJWT